+++ src/expr.h
@@ -7,8 +7,8 @@ using namespace std;
 
 typedef pair<double, string> expr;
 
-/* return expr type from long type */
-expr create_expr (long number);
+/* return expr type from int type */
+expr create_expr (int number);
 
 /* display expr */
 void display_expr (expr var);
